wip work for the crawler/sentiment analysis module update 5

+++ fullon/fullon_daemon.py
@@ -9,7 +9,7 @@ import os
 import signal
 import subprocess
 import time
-from threading import Thread, Event
+from threading import Thread, Event, Lock
 import argparse
 import psutil
 from termcolor import cprint
@@ -18,31 +18,43 @@ from clint.textui import colored
 from libs import log, settings
 from libs.settings_config import fullon_settings_loader  # pylint: disable=unused-import
 from libs.bot_launcher import start as start_launcher, stop as stop_launcher
-from libs.exchange import start_all, stop_all
 from libs.database_ohlcv import start as startohlcv, stop as stopohlcv
 from libs.database import start as startdb, stop as stopdb, WorkerError
+from run.process_manager import ProcessManager
 from run import rpcdaemon_manager as rpc
 from setproctitle import setproctitle
+import sys
 
 logger = log.fullon_logger(__name__)
 thread: Thread
-
-'''
-def handler(signal_received: int, frame) -> None:
-    """ Handle any cleanup here """
-    del signal_received
-    del frame
-    try:
-        thread.join()
-        stop()
-    except AssertionError:
-        pass
-    logger.info('SIGINT or CTRL-C detected. Exiting gracefully')
-    sys.exit()
-'''
+threas2: Thread
+stop_event = Event()
+pmanager = ProcessManager()
+not_stopping = True
 
 
-def kill_processes(name):
+def ignore_signal(sig, frame):
+    process_name = subprocess.check_output(["ps", "-p", str(os.getpid()), "-o", "comm="], encoding="utf-8").strip()
+    if 'Daemon' in process_name:
+        logger.info("SIGINT received again, but ignoring since shutdown is in progress.")
+
+
+def signal_handler(sig, frame):
+    """
+    Cntrl-c shutdown
+    """
+    global not_stopping
+    signal.signal(signal.SIGINT, ignore_signal)
+    process_name = subprocess.check_output(["ps", "-p", str(os.getpid()), "-o", "comm="], encoding="utf-8").strip()
+    if 'Daemon' in process_name:
+        if not_stopping:
+            not_stopping = False
+            logger.warning("Stopping components")
+            stops_components()
+            sys.exit(0)
+
+
+def kill_processes(name, manual: bool = False):
     # Get list of all running processes
     try:
         pids = subprocess.check_output(["pgrep", "-f", name])
@@ -50,7 +62,6 @@ def kill_processes(name):
         return
     if not pids:
         return
-
     pids = pids.splitlines()
     _pids = [int(x.decode('utf-8')) for x in pids]
     pids = sorted(_pids, reverse=True)
@@ -60,33 +71,17 @@ def kill_processes(name):
         pids.remove(my_pid)
     except ValueError:
         pass
+    #import ipdb
+    #ipdb.set_trace()
     for pid in pids:
         try:
             os.kill(int(pid), signal.SIGTERM)
+            if manual:
+                os.kill(int(pid), signal.SIGKILL)
         except ProcessLookupError:
             pass
         except PermissionError:
             logger.error("Permission error: can't kill pid %s", pid)
-    '''
-    time.sleep(1)
-    for pid in pids:
-        try:
-            os.kill(int(pid), signal.SIGKILL)
-        except ProcessLookupError:
-            pass
-    '''
-
-
-def start_pmanager() -> None:
-    """
-    Start the process manager.
-    This function initializes the ProcessManager from the system_manager module,
-    logs a message indicating that the ProcessManager is running, and then starts
-    the process monitor.
-    Returns:
-        None
-    """
-    pass
 
 
 def print_colored_line(char: str, color: str) -> None:
@@ -104,54 +99,23 @@ def print_colored_line(char: str, color: str) -> None:
 
 
 def kill_hard():
+    """
+    Hard kills fullon processes
+    """
     try:
         logger.warning(colored.red("Quitting Fullon... stopping threads..."))
         stop_launcher()
-        stop_all()
         for process in ['Fullon', 'fullon_daemon.py']:
             kill_processes(process)
     except WorkerError:
         pass
-    # pid = os.getpid()
-    # os.kill(pid, signal.SIGTERM)
-    # time.sleep(1)
-    # if os.path.exists(f"/proc/{pid}"):
-    #    os.kill(pid, signal.SIGKILL)
     pass
 
 
-def kill_soft():
-    logger.warning(colored.red("Quitting Fullon... stopping threads..."))
-    rpc.stop_full()
-    thread.join()
-    stop_all()
-    kill_processes('Fullon')
-
-
-def main(cli_args: argparse.Namespace) -> None:
+def print_banners():
     """
-    Main function to manage the execution of the trading bot based on the provided arguments.
-
-    Args:
-        cli_args: argparse.Namespace object containing the parsed command line arguments.
+    Prints banners
     """
-    # Check if the "stop" argument is passed, and stop components accordingly
-    global thread
-
-    if cli_args.stop:
-        kill_processes('Fullon')
-        kill_processes('fullon_daemon')
-        return
-    # Set up signal handling for graceful shutdown on SIGINT or CTRL-C
-    #signal(SIGINT, handler)
-
-    # Remove the 'tmp' directory if it exists and create a new one
-    try:
-        shutil.rmtree('tmp/')
-    except FileNotFoundError:
-        pass
-    os.mkdir('tmp')
-
     # Print the Fullon Crypto Trading Bot title using PyFiglet
     figlet = figlet_format('Fullon Crypto Trading Bot', font='larry3d', width=120)
 
@@ -162,44 +126,100 @@ def main(cli_args: argparse.Namespace) -> None:
     print(colored.yellow(" Fullon Crypto Trading Bot FERI7701027FL9 Crypto Services (C) 2023"))
     print_colored_line('-', 'magenta')
 
-    # Check if an instance of the trading bot is already running and exit if it is
+
+def start_components(cli_args):
+    """
+    Starts fullon components
+    """
+    global thread, stop_event, thread2
+    setproctitle("Fullon Daemon")
+    startohlcv()
+    startdb()
+    start_launcher()
+    setproctitle("Fullon Daemon")
+    # Start the ProcessManager and RPC server
+    thread = Thread(target=rpc.pre_rpc_server,
+                    args=(cli_args.full, cli_args.services, stop_event))
+    thread.start()
+    logger.warning("Console detached")
+    thread2 = Thread(target=pmanager.check_services, args=([stop_event]))
+    thread2.start()
+
+
+def stops_components():
+    """
+    Stops fullon services
+    """
+    global thread, thread2, stop_event
+    stop_event.set()
+    try:
+        thread2.join(timeout=1)
+        if thread and thread.is_alive():
+            thread.join(timeout=5)
+        stop_launcher()
+        stopdb()
+        stopohlcv()
+        kill_hard()
+    except Exception as e:
+        logger.error(f"Error during shutdown: {e}")
+    finally:
+        logger.info("Shutdown complete.")
+
+
+def while_alive():
+    '''
+    Checks whether rpc server thread is alive
+    '''
+    global thread
+    while thread.is_alive():
+        time.sleep(1)
+
+
+def other_instances() -> bool:
+    '''
+    Check if an instance of the trading bot is already running and exit if it is
+    '''
     for process in psutil.process_iter(['pid', 'name', 'username']):
         if "Fullon" in process.name():
             if os.getuid() == process.uids().real:
-                print(colored.red("A daemon is already running, use -x to stop it"))
-                return
-    logger.info("Press CTRL-C to quit")
+                return True
+    return False
+
+
+def main(cli_args: argparse.Namespace) -> None:
+    """
+    Main function to manage the execution of the trading bot based on the provided arguments.
+
+    Args:
+        cli_args: argparse.Namespace object containing the parsed command line arguments.
+    """
+    # Check if the "stop" argument is passed, and stop components accordingly
+
+    if cli_args.stop:
+        kill_processes('Fullon', manual=True)
+        kill_processes('fullon_daemon', manual=True)
+        return
     try:
-        setproctitle("Fullon Daemon")
-        startohlcv()
-        startdb()
-        start_all()
-        start_launcher()
-        stop_event = Event()
-        setproctitle("Fullon Daemon")
-        # Start the ProcessManager and RPC server
-        thread = Thread(target=rpc.pre_rpc_server,
-                        args=(cli_args.full, cli_args.services, stop_event))
-        thread.start()
-        logger.warning("Console detached")
-        while thread.is_alive():
-            time.sleep(2)
+        shutil.rmtree('tmp/')
+    except FileNotFoundError:
+        pass
+    os.mkdir('tmp')
+    print_banners()
+    if other_instances():
+        print(colored.red("A daemon is already running, use -x to stop it"))
+        return
+    logger.info("Press CTRL-C to quit")
+
+    try:
+        start_components(cli_args=cli_args)
+        while_alive()
     except KeyboardInterrupt:
-        try:
-            stop_event.set()
-            stop_launcher()
-            stop_all()
-            stopdb()
-            stopohlcv()
-            time.sleep(1)
-            kill_hard()
-        except (EOFError, KeyboardInterrupt):
-            kill_hard()
-        print("bye...")
-        exit()
+        print("ACCAAAAAAAAAAA")
+        #stops_components()
 
 
 if __name__ == '__main__':
+    signal.signal(signal.SIGINT, signal_handler)
     parser = argparse.ArgumentParser(description="Fullon Daemon Help Page")
     parser.add_argument("-f", "--full", action="store_true",
                         help="Starts all components")
